Remove legacy API based nvJPEG decoder implementation
