+++ cmake/modules/FindNVJPEG.cmake
@@ -59,7 +59,7 @@ if(NVJPEG_FOUND)
   if (NVJPEG_DECOUPLED_API)
     message("nvJPEG is using new API")
   else()
-    message(WARNING " nvJPEG is using the deprecated API")
+    message(FATAL_ERROR "nvjpegBufferPinnedCreate is required but not present in the available version of nvJPEG")
   endif()
 else()
   message("nvJPEG NOT found")
