+++ dali/operators/decoder/nvjpeg/fused/CMakeLists.txt
