+++ dali/operators/decoder/nvjpeg/fused/nvjpeg_decoder_cpu_random_crop.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "dali/operators/decoder/nvjpeg/decoupled_api/fused/nvjpeg_decoder_cpu_random_crop.h"
+#include "dali/operators/decoder/nvjpeg/fused/nvjpeg_decoder_cpu_random_crop.h"
 #include <memory>
 #include <vector>
 #include "dali/util/random_crop_generator.h"
