+++ dali/operators/decoder/nvjpeg/fused/nvjpeg_decoder_cpu_random_crop.h
@@ -12,22 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_RANDOM_CROP_H_
-#define DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_RANDOM_CROP_H_
+#ifndef DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_CPU_RANDOM_CROP_H_
+#define DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_CPU_RANDOM_CROP_H_
 
-#include "dali/operators/decoder/nvjpeg/decoupled_api/nvjpeg_decoder_decoupled_api.h"
+#include "dali/operators/decoder/nvjpeg/nvjpeg_decoder_cpu.h"
 #include "dali/operators/image/crop/random_crop_attr.h"
 
 namespace dali {
 
-class nvJPEGDecoderRandomCrop : public nvJPEGDecoder, public RandomCropAttr {
+class nvJPEGDecoderCPUStageRandomCrop : public nvJPEGDecoderCPUStage, public RandomCropAttr {
  public:
-  explicit nvJPEGDecoderRandomCrop(const OpSpec& spec)
-    : nvJPEGDecoder(spec)
+  explicit nvJPEGDecoderCPUStageRandomCrop(const OpSpec& spec)
+    : nvJPEGDecoderCPUStage(spec)
     , RandomCropAttr(spec)
   {}
 
-  DISABLE_COPY_MOVE_ASSIGN(nvJPEGDecoderRandomCrop);
+  DISABLE_COPY_MOVE_ASSIGN(nvJPEGDecoderCPUStageRandomCrop);
 
  protected:
   CropWindowGenerator GetCropWindowGenerator(int data_idx) const override {
@@ -37,4 +37,4 @@ class nvJPEGDecoderRandomCrop : public nvJPEGDecoder, public RandomCropAttr {
 
 }  // namespace dali
 
-#endif  // DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_RANDOM_CROP_H_
+#endif  // DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_CPU_RANDOM_CROP_H_
