+++ dali/operators/decoder/nvjpeg/fused/nvjpeg_decoder_crop_test.cc
