+++ dali/operators/decoder/nvjpeg/fused/nvjpeg_decoder_random_crop.h
@@ -12,34 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_CROP_H_
-#define DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_CROP_H_
+#ifndef DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_RANDOM_CROP_H_
+#define DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_RANDOM_CROP_H_
 
-#include <vector>
-#include "dali/operators/decoder/nvjpeg/decoupled_api/nvjpeg_decoder_decoupled_api.h"
-#include "dali/operators/image/crop/crop_attr.h"
+#include "dali/operators/decoder/nvjpeg/nvjpeg_decoder_decoupled_api.h"
+#include "dali/operators/image/crop/random_crop_attr.h"
 
 namespace dali {
 
-class nvJPEGDecoderCrop : public nvJPEGDecoder, protected CropAttr {
+class nvJPEGDecoderRandomCrop : public nvJPEGDecoder, public RandomCropAttr {
  public:
-  explicit nvJPEGDecoderCrop(const OpSpec& spec)
+  explicit nvJPEGDecoderRandomCrop(const OpSpec& spec)
     : nvJPEGDecoder(spec)
-    , CropAttr(spec) {
-  }
+    , RandomCropAttr(spec)
+  {}
 
-  DISABLE_COPY_MOVE_ASSIGN(nvJPEGDecoderCrop);
+  DISABLE_COPY_MOVE_ASSIGN(nvJPEGDecoderRandomCrop);
 
  protected:
   CropWindowGenerator GetCropWindowGenerator(int data_idx) const override {
-    return CropAttr::GetCropWindowGenerator(data_idx);
-  }
-
-  void SetupSharedSampleParams(MixedWorkspace &ws) override {
-    CropAttr::ProcessArguments(ws);
+    return RandomCropAttr::GetCropWindowGenerator(data_idx);
   }
 };
 
 }  // namespace dali
 
-#endif  // DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_CROP_H_
+#endif  // DALI_OPERATORS_DECODER_NVJPEG_DECOUPLED_API_FUSED_NVJPEG_DECODER_RANDOM_CROP_H_
