+++ dali/operators/decoder/nvjpeg/nvjpeg_decoder_gpu.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2018, NVIDIA CORPORATION. All rights reserved.
+// Copyright (c) 2019, NVIDIA CORPORATION. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,11 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <string>
-#include "dali/operators/decoder/nvjpeg/legacy_api/nvjpeg_decoder.h"
+#include "dali/operators/decoder/nvjpeg/nvjpeg_decoder_gpu.h"
+#include "dali/pipeline/operator/op_schema.h"
 
 namespace dali {
 
-DALI_REGISTER_OPERATOR(ImageDecoder, nvJPEGDecoder, Mixed);
+DALI_REGISTER_OPERATOR(nvJPEGDecoderGPUStage, nvJPEGDecoderGPUStage, Mixed);
+
+DALI_SCHEMA(nvJPEGDecoderGPUStage)
+  .DocStr(R"code(This operator is the GPU stage of nvJPEGDecoderNew, it is not supposed to be called separately.
+It is automatically inserted during the pipeline creation.)code")
+  .NumInput(3)
+  .NumOutput(1)
+  .MakeInternal()
+  .AddParent("nvJPEGDecoderCPUStage");
 
 }  // namespace dali
