+++ dali/operators/decoder/nvjpeg/nvjpeg_decoder_gpu.h
@@ -20,7 +20,7 @@
 #include <vector>
 
 #include "dali/pipeline/operator/operator.h"
-#include "dali/operators/decoder/nvjpeg/decoupled_api/nvjpeg_helper.h"
+#include "dali/operators/decoder/nvjpeg/nvjpeg_helper.h"
 #include "dali/util/ocv.h"
 #include "dali/core/device_guard.h"
