+++ dali/operators/decoder/nvjpeg/nvjpeg_helper.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include <string>
-#include "dali/operators/decoder/nvjpeg/decoupled_api/nvjpeg_helper.h"
+#include "dali/operators/decoder/nvjpeg/nvjpeg_helper.h"
 
 namespace dali {
